feat(engineioxide/payload): new payload encoder module
Implements v3 & v4 encoder in a custom payload module with different behavior.
Supports binary encoder for v3.
